+++ next/platform/linux/linux.cmake
@@ -31,7 +31,6 @@ target_sources(
         ${MBGL_ROOT}/platform/default/src/mbgl/text/bidi.cpp
         ${MBGL_ROOT}/platform/default/src/mbgl/text/collator.cpp
         ${MBGL_ROOT}/platform/default/src/mbgl/text/local_glyph_rasterizer.cpp
-        ${MBGL_ROOT}/platform/default/src/mbgl/text/unaccent.cpp
         ${MBGL_ROOT}/platform/default/src/mbgl/util/async_task.cpp
         ${MBGL_ROOT}/platform/default/src/mbgl/util/compression.cpp
         ${MBGL_ROOT}/platform/default/src/mbgl/util/format_number.cpp
