+++ next/platform/qt/qt.cmake
@@ -26,7 +26,6 @@ target_sources(
         ${MBGL_ROOT}/platform/default/include/mbgl/gfx/headless_backend.hpp
         ${MBGL_ROOT}/platform/default/include/mbgl/gfx/headless_frontend.hpp
         ${MBGL_ROOT}/platform/default/include/mbgl/gl/headless_backend.hpp
-        ${MBGL_ROOT}/platform/default/include/mbgl/text/unaccent.hpp
         ${MBGL_ROOT}/platform/default/src/mbgl/gfx/headless_backend.cpp
         ${MBGL_ROOT}/platform/default/src/mbgl/gfx/headless_frontend.cpp
         ${MBGL_ROOT}/platform/default/src/mbgl/gl/headless_backend.cpp
@@ -43,7 +42,6 @@ target_sources(
         ${MBGL_ROOT}/platform/default/src/mbgl/storage/online_file_source.cpp
         ${MBGL_ROOT}/platform/default/src/mbgl/storage/sqlite3.cpp
         ${MBGL_ROOT}/platform/default/src/mbgl/text/collator.cpp
-        ${MBGL_ROOT}/platform/default/src/mbgl/text/unaccent.cpp
         ${MBGL_ROOT}/platform/default/src/mbgl/util/compression.cpp
         ${MBGL_ROOT}/platform/qt/src/async_task.cpp
         ${MBGL_ROOT}/platform/qt/src/async_task_impl.hpp
