+++ platform/android/src/text/collator.cpp
@@ -2,8 +2,6 @@
 #include <mbgl/text/language_tag.hpp>
 #include <mbgl/util/platform.hpp>
 
-#include <mbgl/text/unaccent.hpp>
-
 #include <jni/jni.hpp>
 
 #include "../attach_env.hpp"
@@ -144,12 +142,10 @@ public:
         // Because of the difference in locale-awareness, this means turning on case-sensitivity
         // can _potentially_ change compare results for strings that don't actually have any case
         // differences.
-        jni::Local<jni::String> jlhs = jni::Make<jni::String>(*env, useUnaccent ?
-                                                        platform::unaccent(lhs) :
-                                                        lhs);
-        jni::Local<jni::String> jrhs = jni::Make<jni::String>(*env, useUnaccent ?
-                                                        platform::unaccent(rhs) :
-                                                        rhs);
+        jni::Local<jni::String> jlhs = useUnaccent ? android::StringUtils::unaccent(*env, jni::Make<jni::String>(*env, lhs))
+                                                     : jni::Make<jni::String>(*env, lhs);
+        jni::Local<jni::String> jrhs = useUnaccent ? android::StringUtils::unaccent(*env, jni::Make<jni::String>(*env, rhs))
+                                                     : jni::Make<jni::String>(*env, rhs);
 
         jni::jint result = android::Collator::compare(*env, collator, jlhs, jrhs);
